Update CI to use Node 12
The CONTRIBUTING.md file states that you should "make sure your node,
npm, and yarn are on the latest versions", however the CI uses an
older Node version.

This PR updates the CI to use the latest Node version (12), in order
to align the CI with the statement in the documentation.